pyproject.toml
[tool.poetry]
name = "yuzu-compat-bot"
version = "1.4.10"
description = "A discord bot to maintain a list of games compatible with yuzu."
authors = ["Larson Ashcroft <larsona@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
rich = "^10.7.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[tool.autopep8]
max_line_length = 155

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
